feat(events): unify and clean up types
