+++ packages/commons/src/events/Form.ts
@@ -1,10 +1,9 @@
 import { NonEmptyArray } from 'src/types'
 import { z } from 'zod'
-import { Field, Label } from '../utils'
-import { SystemRoleType, systemRoleTypes } from './ActionConfig'
+import { Field, Label, SystemRoleType, systemRoleTypes } from './utils'
 
 export const FormGroupField = Field.extend({
-  id: z.string(), // @TODO: Check whether we should always have id.
+  id: z.string(),
   type: z.string(), // @TODO: Get enums from somewhere, field types
   required: z.boolean(),
   searchable: z.boolean(),
@@ -13,7 +12,6 @@ export const FormGroupField = Field.extend({
 
 export const FormSection = z.object({
   title: Label,
-  // Should this be fields
   groups: z.array(FormGroupField)
 })
 
@@ -23,8 +21,6 @@ export const Form = z.object({
     id: z.string(),
     label: Label
   }),
-  // actions: z.array(z.string()), // @TODO: ignore
   roles: z.array(z.enum(systemRoleTypes as NonEmptyArray<SystemRoleType>)),
   form: z.array(FormSection)
-  // review: z.array(ReviewSection) // Review will be brought when needed.
 })
